refactor: note 관련 query 메서드를 entities/note/api/noteQueries.ts로 통합
